fix some issues with adding multiple products to the cart

+++ view/assets/javascript/shop/shop_productPage.js
@@ -9,22 +9,41 @@ if(cartButton !== null) {
         if(localStorageData === null) {
             localStorageData = []
         }
-       
-        let data = {
-            'product_id' : cartButton.dataset.id,
-            'product_name': cartButton.dataset.name,
-            'product_description': cartButton.dataset.description,
-            'product_image': cartButton.dataset.image,
-            'product_price': cartButton.dataset.price,
-            'product_quantity': cartButton.dataset.quantity
-        }
+
+        // Check if the product is already in the list
+        const existingProductIndex = localStorageData.findIndex(storedProduct => storedProduct.product_id === cartButton.dataset.id);
+
+        
+
+        if(existingProductIndex !== -1) {
+            // The product is already in the list, so just increase its quantity
+            localStorageData[existingProductIndex].product_quantity++;
+
+            window.localStorage.setItem('products-cart-items', JSON.stringify(localStorageData));
+
+        }else {
+            let data = {
+                'product_id' : cartButton.dataset.id,
+                'product_name': cartButton.dataset.name,
+                'product_description': cartButton.dataset.description,
+                'product_image': cartButton.dataset.image,
+                'product_price': cartButton.dataset.price,
+                'product_quantity': cartButton.dataset.quantity
+            }
+            localStorageData.push(data);
+            window.localStorage.setItem('products-cart-items', JSON.stringify(localStorageData))
+       }
+        
             
          
-        localStorageData.push(data);
 
-        window.localStorage.setItem('products-cart-items', JSON.stringify(localStorageData))
+
+
+        location.reload();
+        
     });
 }
+
 
 const products = JSON.parse(localStorage.getItem('products-cart-items'));
 const productId = document.getElementById('id').value;
@@ -46,12 +65,12 @@ const ProductQuantity = Number(ObjectFromLocalStorageById(productId).product_qua
 
 const totalPriceContainer = document.getElementById('order-informations');
 
-
 totalPriceContainer.innerHTML = `
     <span class="text-xl font-bold text-slate-600 mx-10 w-[50%]">Product summary</span>
     <span class="text-lg font-bold text-slate-500 mx-10 w-[50%]">Total price : <span class="totalPrice">${ProductPrice * ProductQuantity}</span></span>
-    
-    `
+
+`
+
 
 
 // loop over plus and minus buttons and add event listener to them 
@@ -60,7 +79,6 @@ const plus_button = document.querySelectorAll("#plusButton");
 const minus_button = document.querySelectorAll("#minusButton");
 const cancel_button = document.querySelectorAll("#cancelButton");
 const total_price = document.getElementById('totalPrice');
-
 
 for(let plus of plus_button){
     plus.addEventListener("click" , _ => {
@@ -82,10 +100,16 @@ for(let plus of plus_button){
             'product_quantity' : plus.nextElementSibling.dataset.quantity
         }
 
-        const index = localStorageData.findIndex(item => data.product_id === Product.product_id);
+        const index = localStorageData.findIndex(data => data.product_id === Product.product_id);
         localStorageData.splice(index, 1 , data);
 
         window.localStorage.setItem('products-cart-items', JSON.stringify(localStorageData));
+
+        totalPriceContainer.innerHTML = `
+            <span class="text-xl font-bold text-slate-600 mx-10 w-[50%]">Product summary</span>
+            <span class="text-lg font-bold text-slate-500 mx-10 w-[50%]">Total price : <span class="totalPrice">${ProductPrice * ProductQuantity}</span></span>
+    
+        `
 
 
     })
@@ -111,10 +135,17 @@ for(let minus of minus_button){
             'product_quantity' : minus.previousElementSibling.dataset.quantity
         }
 
-        const index = localStorageData.findIndex(item => data.product_id === Product.product_id);
+        const index = localStorageData.findIndex(data => data.product_id === Product.product_id);
         localStorageData.splice(index, 1 , data);
 
         window.localStorage.setItem('products-cart-items', JSON.stringify(localStorageData));
+
+
+        totalPriceContainer.innerHTML = `
+            <span class="text-xl font-bold text-slate-600 mx-10 w-[50%]">Product summary</span>
+        <span class="text-lg font-bold text-slate-500 mx-10 w-[50%]">Total price : <span class="totalPrice">${ProductPrice * ProductQuantity}</span></span>
+    
+        `
 
 
     })
